requirements.txt
bcrypt==4.0.1
Flask==2.3.2
Flask-Cors==4.0.0
PyJWT==2.8.0
pymysql==1.1.0
requests==2.31.0
openai==0.28.0
Flask-MySQL==1.5.2
python-dotenv==1.0.0

